--- my-app/src/App.css
@@ -1,237 +1,0 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-/* About */
-.about {
-  margin: 100px 0 150px 0;
-  display: flex;
-  justify-content: center;
-  gap: 200px;
-}
-.about__info {
-  display: flex;
-  flex-direction: column;
-  max-width: 400px;
-  align-items: center;
-  gap: 20px;
-}
-.about__title_light {
-  font-weight: 700;
-  font-size: 58px;
-  line-height: 70px;
-  letter-spacing: -1px;
-  color: black;
-  transition: all 0.1s ease-in-out;
-}
-.about__title_light:hover {
-  background-color: black;
-  color: white;
-}
-
-.about__title_dark {
-  font-weight: 700;
-  font-size: 58px;
-  line-height: 70px;
-  letter-spacing: -1px;
-  color: #e8f000;
-  transition: all 0.1s ease-in-out;
-}
-.about__title_dark:hover {
-  background-color: #e8f000;
-  color: #3a3a3a;
-}
-
-.about__text_light {
-  color: black;
-  font-weight: 500;
-  font-size: 22px;
-  line-height: 28px;
-  letter-spacing: 0px;
-  padding: 20px 0 100px 0;
-}
-
-.about__text_dark {
-  color: #e0e0ff;
-  font-weight: 500;
-  font-size: 22px;
-  line-height: 28px;
-  letter-spacing: 0px;
-  padding: 20px 0 100px 0;
-}
-
-.about__link_light {
-  width: fit-content;
-  padding: 17px 27px;
-  border: none;
-  background-color: rgb(181, 241, 0);
-  color: rgb(58, 58, 58);
-  border-radius: 100px;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 26px;
-  transition: all 0.3s ease-in-out;
-}
-.about__link_light:hover {
-  background-color: rgb(234, 255, 172);
-  cursor: pointer;
-}
-.about__link_dark {
-  width: fit-content;
-  padding: 17px 27px;
-  border: none;
-  color: #3a3a3a;
-  background-color: #e8f000;
-
-  border-radius: 100px;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 26px;
-  transition: all 0.3s ease-in-out;
-}
-.about__link_dark:hover {
-  color: #999e00;
-  background-color: #e8ed5e;
-  cursor: pointer;
-}
-
-/* Error */
-.error__container {
-  display: flex;
-  flex-direction: column;
-  max-width: 400px;
-  align-items: center;
-  gap: 20px;
-}
-
-.error_title {
-  font-weight: 700;
-  font-size: 58px;
-  line-height: 70px;
-  letter-spacing: -1px;
-  color: black;
-  transition: all 0.1s ease-in-out;
-}
-.error_massage {
-  font-size: 40px;
-  letter-spacing: -1px;
-  color: black;
-}
-
-/* Navigation */
-.nav {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  top: 0;
-  padding-top: 30px;
-}
-.nav__logo {
-  width: 5em;
-  transition: transform 0.3s ease;
-}
-.nav__logo:hover {
-  transform: scale(1.1);
-}
-
-.nav__List {
-  display: flex;
-  align-items: center;
-  gap: 20px;
-}
-
-a {
-  font-weight: 500;
-  text-decoration: inherit;
-}
-
-.nav__link_light {
-  color: #333;
-
-  transition: all 0.3s ease;
-}
-.nav__link_light:hover {
-  color: #9c8000;
-}
-.nav__link_dark {
-  color: #e8f000;
-
-  transition: all 0.3s ease;
-}
-.nav__link_dark:hover {
-  color: #999e00;
-}
-
-/* PokemonCard an Details */
-.cardInfoLight {
-  font-size: 1rem;
-  color: #333;
-  font-weight: 500;
-}
-.cardInfoDark {
-  color: #e0e0ff;
-  font-size: 1rem;
-  font-weight: 500;
-}
-
-.cardTitleLight {
-  font-size: 1.2rem;
-  color: #333;
-  font-weight: 500;
-}
-.cardTitleDark {
-  color: #e0e0ff;
-  font-size: 1.2rem;
-  font-weight: 500;
-}
-
-/* Footer */
-footer {
-  font-size: 1.2rem;
-  font-weight: 500;
-}
-.footerLight {
-  color: #333;
-}
-.footerDark {
-  color: #e0e0ff;
-}
-
-.button__pagination_light {
-  color: #333;
-
-  background-color: #f0eded;
-
-  transition: all 0.3s ease;
-}
-.button__pagination_light:hover {
-  color: #9c8000;
-
-  background-color: #d8db82;
-
-  transition: all 0.3s ease;
-}
-.button__pagination_dark {
-  color: #3a3a3a;
-  background-color: #e8f000;
-
-  transition: all 0.3s ease;
-}
-.button__pagination_dark:hover {
-  color: #999e00;
-  background-color: #e8ed5e;
-}
-.textLight {
-  font-size: 1rem;
-  color: #333;
-  font-weight: 500;
-}
-.textDark {
-  color: #e0e0ff;
-  font-size: 1rem;
-  font-weight: 500;
-}
