+++ my-app/src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import { Navbar } from './NavBar';
-import { useTheme } from './ThemeContext';
+import { Navbar } from './NavBar/NavBar';
+import { useTheme } from './ThemeContext/ThemeContext';
 
 export const Layout = () => {
   const { theme } = useTheme();
