+++ my-app/src/components/PokemonItem/PokemonItem.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styles from './PokemonItem.module.css';
-import { useTheme } from '../ThemeContext';
+import { useTheme } from '../ThemeContext/ThemeContext';
 
 export const PokemonItem = (props: { name: string }) => {
   const { theme } = useTheme();
