+++ my-app/src/main.tsx
@@ -3,8 +3,7 @@ import { createRoot } from 'react-dom/client';
 import './index.css';
 import { App } from './App.tsx';
 import { BrowserRouter } from 'react-router-dom';
-import { ThemeProvider } from './components/ThemeContext.tsx';
-import './style/variable.css';
+import { ThemeProvider } from './components/ThemeContext/ThemeContext.tsx';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
