--- my-app/src/style/variable.css
@@ -1,2 +1,0 @@
-:root {
-}
